requirements.txt
setuptools~=57.0.0
requests~=2.26.0
python-dotenv~=0.19.2
beautifulsoup4~=4.10.0
numpy~=1.21.4
Pillow~=9.0.1
Flask~=2.2.2
PyYAML~=6.0
pandas~=1.3.5
pywin32~=305
lxml~=4.7.1
pytz~=2021.3